pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aura-ctx"
version = "0.1.0"
description = "Your portable AI context. Carry your identity across every AI tool."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Nok (Enoch Afanwoubo)", email = "nok@distalinc.com" },
]
keywords = ["ai", "context", "memory", "mcp", "portable", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
extract = [
    "openai>=1.0",
    "tiktoken>=0.5",
]
mcp = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
aura = "aura.cli:app"

[project.urls]
Homepage = "https://github.com/distalinc/aura"
Documentation = "https://aura.distalinc.com"
Repository = "https://github.com/distalinc/aura"
Issues = "https://github.com/distalinc/aura/issues"

[tool.hatch.build.targets.wheel]
packages = ["aura"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
